types split to folders
